Spring: Use POST for sign out
This reverts the effect of 22b70d442332b64439dced62c6fd11813ff1a69b.
I just discovered that the project's plan is to use POST for the
sign out. This is also the conventional method for Spring security,
as it protects the sign out against CSRF.

Resolves: komarserjio/notejam#45
See also: komarserjio/notejam#52