Spring: Implement password recovery token
When the user requests a password recovery, a secure random token
with an expiration date is stored in the database. Expired tokens
are purged from the database on a daily schedule.

The process does silently stop if the email address is not
associated to an existing account. This is done on purpose to not
reveal information about existing users.

See also: komarserjio/notejam#52